Random generation and Duplicate Protection
We can now spawn a dev-defined number of prefabs at start and in a random order; there is duplicate protection to ensure that there are at least two different prefabs between each occurence of a prefab (i.e. you will not see the same chunk twice in a row or even with one other chunk in between). This means we will need an aboslute minumum of 3 prefab chunks (although this would result in a randomly seeded pattern), ideally 4 or more unique prefabs.